Move openssl_privatekey_info code to module_utils.
